调试升级，dac输出还有些问题

+++ User/inc/delay.h
@@ -3,9 +3,9 @@
 
 #include "types.h"
 
-void delay_us(u16 us);
+void delay_us(u32 us);
 
-void delay_ms(u16 ms);
+void delay_ms(u32 ms);
 
 #endif
